Update README.md
The Notebook is a story of unwavering love that has gathered a following of fans. Additionally, this film depicts the effects of dementia on patients as well as their caregivers. This movie conveys its message through a love story, yet highlights more grave issues related to dementia through subtler means.